extract domain.go
